% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{benfordfct}
\alias{benfordfct}
\title{Benford function}
\usage{
benfordfct(data,digit)
}
\arguments{
\item{data}{one column dataset with type integer.}

\item{digit}{number within this interval [1;9] & "12"}
}
\value{
A list with confidence interval, chi2 stats, ssd & zstat
}
\description{
It gets all the summary and statistics for a dataset.
\code{\link{benford}}.
}
\details{
Benford Ely for benford introduction and example analytics

The Benford Ely package provides an easy function to compute
all the basic summary and statistics that you need to perform
an analysis at your data with regard to Benford Law.
And then identify data that could have been manipulated and which
will need further verification.

More information can be found on its help documentation.

The main and only function is \code{\link{benfordfct}}.

The package prints confidence intervals, chi2 stats, z and ssd values.


The package also includes a complete study analysing the US 2020 elections

##' @examples
benfordfct(data,digit) # generates all the statistics mentionned above
data is a column dataset number
digit is by default 1
digit can be in the interval [1:9] which represent its digit
and digit can be "12" which represent first-two digits
}
\examples{
data(data) #gets data
benfordfct(data)
}
