% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.functions.R
\name{chisq}
\alias{chisq}
\title{Benford function}
\usage{
benfordfct(data,digit)
}
\arguments{
\item{data}{one column dataset with type integer.}

\item{digit}{number within this interval [1;9] & "12"}
}
\value{
A list with confidence interval, chi2 stats, ssd & zstat
}
\description{
It gets all the summary and statistics for a dataset.
\code{\link{benford}}.
}
\examples{
data(data) #gets data
benfordfct(data)
}
